+++ docs/src/changelog.md
@@ -12,6 +12,12 @@ If your currently installed *Quality-time* version is not the latest version, pl
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
 
+## [Unreleased]
+
+### Added
+
+- When measuring 'slow transactions' using Grafana k6 as source, allow for filtering transactions and configuring the response times to evaluate. Closes [#11462](https://github.com/ICTU/quality-time/issues/11462).
+
 ## v5.33.0 - 2025-06-20
 
 ### Changed